Move plugin into plugin.ts
