+++ src/index.ts
@@ -1,53 +1,9 @@
-import componentsReturnOnce from "./rules/components-return-once";
-import eventHandlers from "./rules/event-handlers";
-import imports from "./rules/imports";
-import jsxNoDuplicateProps from "./rules/jsx-no-duplicate-props";
-import jsxNoScriptUrl from "./rules/jsx-no-script-url";
-import jsxNoUndef from "./rules/jsx-no-undef";
-import jsxUsesVars from "./rules/jsx-uses-vars";
-import noDestructure from "./rules/no-destructure";
-import noInnerHTML from "./rules/no-innerhtml";
-import noProxyApis from "./rules/no-proxy-apis";
-import noReactDeps from "./rules/no-react-deps";
-import noReactSpecificProps from "./rules/no-react-specific-props";
-import noUnknownNamespaces from "./rules/no-unknown-namespaces";
-import preferClasslist from "./rules/prefer-classlist";
-import preferFor from "./rules/prefer-for";
-import preferShow from "./rules/prefer-show";
-import reactivity from "./rules/reactivity";
-import selfClosingComp from "./rules/self-closing-comp";
-import styleProp from "./rules/style-prop";
-import noArrayHandlers from "./rules/no-array-handlers";
-// import validateJsxNesting from "./rules/validate-jsx-nesting";
+import { type AllRules, plugin } from "./plugin";
 
-const allRules = {
-  "components-return-once": componentsReturnOnce,
-  "event-handlers": eventHandlers,
-  imports,
-  "jsx-no-duplicate-props": jsxNoDuplicateProps,
-  "jsx-no-undef": jsxNoUndef,
-  "jsx-no-script-url": jsxNoScriptUrl,
-  "jsx-uses-vars": jsxUsesVars,
-  "no-destructure": noDestructure,
-  "no-innerhtml": noInnerHTML,
-  "no-proxy-apis": noProxyApis,
-  "no-react-deps": noReactDeps,
-  "no-react-specific-props": noReactSpecificProps,
-  "no-unknown-namespaces": noUnknownNamespaces,
-  "prefer-classlist": preferClasslist,
-  "prefer-for": preferFor,
-  "prefer-show": preferShow,
-  reactivity,
-  "self-closing-comp": selfClosingComp,
-  "style-prop": styleProp,
-  "no-array-handlers": noArrayHandlers,
-  // "validate-jsx-nesting": validateJsxNesting
-};
+export type { AllRules };
 
-export type AllRules = keyof typeof allRules;
-
-const plugin = {
-  rules: allRules,
+const pluginLegacy = {
+  rules: plugin.rules,
   configs: {
     recommended: {
       plugins: ["solid"],
@@ -132,6 +88,6 @@ const plugin = {
   },
 };
 // Must be module.exports for eslint to load everything
-module.exports = plugin;
+module.exports = pluginLegacy;
 
-export type Configs = keyof typeof plugin.configs;
+export type Configs = keyof typeof pluginLegacy.configs;
